refactor(admin-content): enable layz loading for Events/Pages/Posts
