+++ modules/admin-content/src/content-routing.module.ts
@@ -3,10 +3,6 @@ import { Routes, RouterModule } from '@angular/router'
 
 import { ContentDashboardComponent } from './content.component'
 
-import { PostsRoutes } from './posts/posts.routes'
-import { PagesRoutes } from './pages/pages.routes'
-import { EventsRoutes } from './events/events.routes'
-
 import { FilesComponent } from './files/files.component'
 import { ProductsComponent } from './products/products.component'
 
@@ -23,10 +19,10 @@ export const routes: Routes = [
     resolve: { domain: DomainResolver },
     children: [
       { path: '', component: ContentDashboardComponent, data: { title: 'Dashboard' } },
+      { path: '', loadChildren: './events/events.module#EventsModule' },
+      { path: '', loadChildren: './pages/pages.module#PagesModule' },
+      { path: '', loadChildren: './posts/posts.module#PostsModule' },
       { path: 'files', component: FilesComponent, data: { title: 'Files' } },
-      ...PagesRoutes,
-      ...PostsRoutes,
-      ...EventsRoutes,
       { path: 'products', component: ProductsComponent, data: { title: 'Products' } },
     ],
   },
