+++ modules/admin-content/src/pages/components/page-detail.component.ts
