+++ modules/admin-content/src/pages/components/page-list.component.ts
